requirements.txt
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
concurrent-log-handler==0.9.25
contourpy==1.3.1
cycler==0.12.1
datasets==3.4.1
dill==0.3.8
emoji==1.7.0
et_xmlfile==2.0.0
evaluate==0.4.3
faiss-cpu==1.10.0
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
huggingface-hub==0.29.3
idna==3.10
imbalanced-learn==0.13.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lightgbm==4.6.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numpy==2.1.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
peft==0.15.0
pillow==11.0.0
portalocker==3.1.1
propcache==0.3.0
psutil==7.0.0
pyarrow==19.0.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==3.4.1
setuptools==77.0.3
six==1.17.0
sklearn-compat==0.1.3
sympy==1.13.3
text2emotion==0.0.5
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.8.0.dev20250327+cu118
torchaudio==2.6.0.dev20250329+cu118
torchvision==0.22.0.dev20250329+cu118
tqdm==4.67.1
transformers==4.50.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1
wordcloud==1.9.4
xgboost==3.0.0
xxhash==3.5.0
yarl==1.18.3
